Extract legacy unit tests
